+++ test/artifacts_test.ts
@@ -27,7 +27,7 @@ describe('Artifacts', () => {
             await (zeroEx.token as any)._getProxyAddressAsync();
         }).timeout(TIMEOUT);
         it('exchange contract is deployed', async () => {
-            const exchangeContractAddresses = await zeroEx.exchange.getAvailableContractAddressesAsync();
+            const exchangeContractAddresses = await zeroEx.getAvailableExchangeContractAddressesAsync();
             expect(exchangeContractAddresses).to.have.lengthOf.above(0);
         }).timeout(TIMEOUT);
     });
